backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.20.0
greenlet>=3.0.0
alembic>=1.13.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
redis>=5.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
scikit-learn>=1.4.0
numpy>=1.26.0
